pyproject.toml
# Copyright 2026 Enactic, Inc.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openarm_ker"
description = "Teleoperation system for OpenArm robots using KER devices"
version = "0.1.0"
authors = [{name = "Enactic, Inc."}]
license = {text = "Apache-2.0"}
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "numpy>=1.20.0",
    "PyYAML>=5.4.0",
    "pyserial",
    "openarm-can>=0.1.0",
]

[project.urls]
changelog = "https://github.com/enactic/openarm_ker/releases"
issues = "https://github.com/enactic/openarm_ker/issues"
repository = "https://github.com/enactic/openarm_ker.git"

[dependency-groups]
dev = [
    "ruff == 0.14.11", # sync with .pre-commit-config.yaml
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
openarm_ker= ["config/*.yaml"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff.lint]
extend-select = [
  "D",   # pydocstyle
  "UP",
]

[tool.ruff.lint.per-file-ignores]
"!src/**.py" = ["D"]
